Removes models folder and restructures some model structs
